AxiToAxiLitePc: Fix bug with multiple parallel transactions
Before this fix, when multiple requests arrived at the input from the
upstream AXI bus, a new transaction on the downstream AxiLite bus could
be started before the previous transaction was finished.  This led to
write addresses not getting registered and thus data being sent to wrong
locations.

This update enforces that a new address from the upstream AXI bus is
only accepted when the downstream transaction is completed.