AxiToAxiLitePc: Ensure compatibility with non-Xilinx slaves
The previous commit assumed that the AW and the W handshake would occur
in the same cycle.  This behavior is specific to Xilinx slaves and not
generally true for AXI-compliant slaves.

This commit removes that assumption and makes this module work correctly
also if the AW and the W handshake occur in different cycles.  To make
sure that the Write Response handshake does not go by unnoticed, the
`b_valid` and `b_ready` are now only active in the corresponding state.